use Beta module
