+++ static/arguments.yml
@@ -132,6 +132,13 @@
     type: Hash
     default: N/A
     description: "[System arguments](/system-arguments)"
+- component: Breadcrumbs
+  source: https://github.com/primer/view_components/tree/main/app/components/primer/beta/breadcrumbs.rb
+  parameters:
+  - name: system_arguments
+    type: Hash
+    default: N/A
+    description: "[System arguments](/system-arguments)"
 - component: Text
   source: https://github.com/primer/view_components/tree/main/app/components/primer/beta/text.rb
   parameters:
@@ -230,13 +237,6 @@
     description: "[System arguments](/system-arguments)"
 - component: Box
   source: https://github.com/primer/view_components/tree/main/app/components/primer/box_component.rb
-  parameters:
-  - name: system_arguments
-    type: Hash
-    default: N/A
-    description: "[System arguments](/system-arguments)"
-- component: Breadcrumbs
-  source: https://github.com/primer/view_components/tree/main/app/components/primer/breadcrumbs.rb
   parameters:
   - name: system_arguments
     type: Hash
