+++ static/statuses.json
@@ -7,14 +7,14 @@
   "Primer::Beta::AutoComplete::Item": "beta",
   "Primer::Beta::Avatar": "beta",
   "Primer::Beta::AvatarStack": "beta",
+  "Primer::Beta::Breadcrumbs": "beta",
+  "Primer::Beta::Breadcrumbs::ItemComponent": "alpha",
   "Primer::Beta::Text": "beta",
   "Primer::Beta::Truncate": "beta",
   "Primer::Beta::Truncate::TruncateText": "alpha",
   "Primer::BlankslateComponent": "beta",
   "Primer::BorderBoxComponent": "beta",
   "Primer::BoxComponent": "stable",
-  "Primer::Breadcrumbs": "beta",
-  "Primer::Breadcrumbs::ItemComponent": "alpha",
   "Primer::ButtonComponent": "beta",
   "Primer::ButtonGroup": "beta",
   "Primer::ClipboardCopy": "beta",
